add the documentation
Now that I think of it, deriving the nonlinear controller using Lyapunov analysis and then linearizing it defeats the purpose of the stability analysis? Not sure (?). If I had wanted a linear controller, I should have linearized the dynamics first and then performed the stability analysis and control design accordingly.